requirements.txt
Django
djangorestframework
django-rest-passwordreset
psycopg2-binary
drf-spectacular